changed functions from a previous commit
