+++ internal/service/admin/refreshAllData.go
@@ -2,8 +2,6 @@ package admin
 
 import (
 	"context"
-	"homeworktodolist/internal/entity"
-	"homeworktodolist/internal/utils/classParse"
 )
 
 // TODO: дописать метод сделать транзакцию и удаление предыдущих
@@ -13,22 +11,13 @@ func (s *Service) RefreshAllData(ctx context.Context) error {
 	if err != nil {
 		return err
 	}
+
 	for _, group := range groups {
-		class, err := classParse.IcalParse(group.IcalLink)
+		err = s.subjectService.UpdGroupSubjects(ctx, group)
 		if err != nil {
 			return err
 		}
-		subjects := classParse.ParseSubject(class)
-		for _, subject := range subjects {
-			err = s.subjectService.Create(ctx, entity.Subject{
-				GroupId:     group.GroupID,
-				SubjectName: subject,
-			})
-			if err != nil {
-				return err
-			}
-		}
-
 	}
+
 	return nil
 }
